requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
Flask_SQLAlchemy==3.1.1
pyodbc==5.3.0
python_dotenv==1.0.1
bcrypt==4.1.3
SQLAlchemy==2.0.30
Werkzeug==3.0.3